refactoring helpers into its own file
